Extract and reuse caching logic
