+++ src/configuration.js
@@ -1,7 +1,6 @@
 import { getStorageItem, setStorageItem } from "./browser";
 
 const CONFIG_KEY = "ld_ext_config";
-const TAB_METADATA_CACHE_KEY = "ld_tab_metadata_cache";
 const DEFAULTS = {
   baseUrl: "",
   token: "",
@@ -20,20 +19,6 @@ export async function getConfiguration() {
 export async function saveConfiguration(config) {
   const configJson = JSON.stringify(config);
   await setStorageItem(CONFIG_KEY, configJson);
-}
-
-export async function getCachedTabMetadata() {
-  const json = await getStorageItem(TAB_METADATA_CACHE_KEY);
-  return json ? JSON.parse(json) : null;
-}
-
-export async function cacheTabMetadata(tabMetadata) {
-  const json = JSON.stringify(tabMetadata);
-  await setStorageItem(TAB_METADATA_CACHE_KEY, json);
-}
-
-export async function clearCachedTabMetadata() {
-  await cacheTabMetadata(null);
 }
 
 export function isConfigurationComplete(config) {
